# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pro Disciplinas",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost/api"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Athentication"
                ],
                "summary": "Login",
                "description": "Route to login to the application",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "List of pre-offer"
                    }
                }
            }
        },
        "/me": {
            "post": {
                "tags": [
                    "Athentication"
                ],
                "summary": "Who i'm?",
                "description": "Route to return the logged in user",
                "operationId": "af3b8d149797532336b3d65ef346c917",
                "responses": {
                    "200": {
                        "description": "A User"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Athentication"
                ],
                "summary": "Logout",
                "description": "Make the token invalid",
                "operationId": "68a6108af04af5e3bc1ea68a6c6f1299",
                "responses": {
                    "200": {
                        "description": "A User"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/refresh": {
            "post": {
                "tags": [
                    "Athentication"
                ],
                "summary": "Generate a new token",
                "description": "Generate a new token",
                "operationId": "0c3ea3c61aed881016646810460499e9",
                "responses": {
                    "200": {
                        "description": "A User"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/period": {
            "get": {
                "tags": [
                    "/period"
                ],
                "summary": "Display a listing of the resource",
                "description": "Get all period on database",
                "operationId": "43abf9c9365c4a309078c55423c3cc1b",
                "responses": {
                    "200": {
                        "description": "List of period"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "/period"
                ],
                "summary": "Store a newly created resource in storage.",
                "description": "Post period on database",
                "operationId": "8f6be959f5053f09e0a98105993b046c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "start_request": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "end_request": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "start_indication": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "end_indication": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "List of period"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/period/{id}": {
            "get": {
                "tags": [
                    "/period"
                ],
                "summary": "Display a specified resource",
                "description": "Show period",
                "operationId": "342880a73e5934553cd8cdc250c5fc19",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Period id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show period"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "/period"
                ],
                "summary": "Update the specified resource in storage.",
                "description": "Update period on database",
                "operationId": "5365b76611b541bd4c18166a47d09753",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Period id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "start_request": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "end_request": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "start_indication": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "end_indication": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update of period"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "/period"
                ],
                "summary": " Remove the specified resource from storage.",
                "description": "Remove a period on database",
                "operationId": "d673774e1b87d7290a563efc2debcb1b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Period Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "period deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}